venus-image: shrink the size
Commit 245cfbee7fa65 "nanopi: enlarge filesystem to the partition size", makes
the filesystem as large as the partition. However this is done for ext by padding
with zeros and hence swupdate ends up writting all these zero to flash. This takes
long and cause unnecessary writes. So don't do this, and developers wanting more
space can now simply call:

/opt/victronenergy/swupdate-scripts/resize2fs.sh

on their device, to enlarge the rootfs.

Keeping the rpi for now, since it uses the size of the image also for the layout of
the sdcard it is running from.